Quant_Unified/requirements.txt
aiodns==3.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
beautifulsoup4==4.14.3
bleach==6.3.0
ccxt==4.5.26
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
dataframe_image==0.2.7
defusedxml==0.7.1
ecdsa==0.19.1
fastapi==0.124.2
fastjsonschema==2.21.2
fonttools==4.61.0
frozenlist==1.8.0
greenlet==3.3.0
h11==0.16.0
h5py==3.15.1
hdf5plugin==6.0.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.9
llvmlite==0.46.0
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
mistune==3.1.4
more-itertools==10.8.0
multidict==6.7.0
narwhals==2.13.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numba==0.63.1
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
pillow==12.0.0
platformdirs==4.5.1
playwright==1.57.0
plotly==6.5.0
propcache==0.4.1
pyasn1==0.6.1
pycares==4.11.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
pyotp==2.9.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
python-socks==2.8.0
pytz==2025.2
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.45
starlette==0.50.0
tinycss2==1.4.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
uvicorn==0.38.0
webencodings==0.5.1
websockets==15.0.1
yarl==1.22.0
